requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
deepgram-sdk==3.7.0
twilio==8.10.0
websockets==12.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
requests==2.31.0
pydub==0.25.1
gevent==23.9.1
flask-socketio==5.3.6
eventlet==0.33.3